requirements.txt
python-telegram-bot==20.7
yt-dlp
ffmpeg-python
python-dotenv
